requirements.txt
asgiref>=3.3.1
Pillow>=8.1.1
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
whitenoise>=5.2.0
gunicorn
python-dotenv>=0.15.0

djangorestframework>=3.12.2
Django>=3.1.6
django-crispy-forms>=1.11.1
django-heroku>=0.3.1
django-money>=1.3.1
django-richtextfield>=1.6
dj-database-url>=0.5.0
django-filter>=2.4.0
django-imagekit>=4.0.2

#dev
black==20.8b1
isort==5.8.0